requirements.txt
python-telegram-bot==13.15
gTTS
Pillow
pandas
numpy
python-dotenv
python-telegram-bot-calendar
newrelic
requests
pytz
google-api-python-client
google-auth-oauthlib
docx2pdf
telegram
unidecode
python-decouple
pytz
psycopg2
sqlalchemy